fixing bugs

+++ delete/deleteMessage.go
@@ -12,8 +12,10 @@ func DeleteMessage(r *http.Request, db *sql.DB) {
 		fmt.Println("test")
 		var owner string
 		var uuidPath string
+		var firstMessage string
+		var message string
 
-		query := fmt.Sprintf("SELECT owner, uuidPath FROM messages WHERE uuid = '%s'", r.FormValue("delete"))
+		query := fmt.Sprintf("SELECT owner, uuidPath, message FROM messages WHERE uuid = '%s'", r.FormValue("delete"))
 
 		row, err := db.Query(query)
 		if err != nil {
@@ -21,22 +23,43 @@ func DeleteMessage(r *http.Request, db *sql.DB) {
 		} else {
 			for row.Next() {
 				defer row.Close()
-				err = row.Scan(&owner, &uuidPath)
+				err = row.Scan(&owner, &uuidPath, &message)
 				if err != nil {
 					fmt.Println(err)
 				}
 			}
-			if owner == t.USER.Username || t.USER.Admin == 1 {
-				query2 := fmt.Sprintf("DELETE FROM messages WHERE uuid = '%s'", r.FormValue("delete"))
-				db.Exec(query2)
-				query = fmt.Sprintf("DELETE FROM likesFromUser WHERE uuidLiked = '%s'", r.FormValue("delete"))
-				db.Exec(query)
-				fmt.Println("message deleted")
-				query3 := fmt.Sprintf("UPDATE topics SET nmbPosts = nmbPosts - 1 WHERE uuid = '%s'", uuidPath)
-				db.Exec(query3)
 
+			query2 := fmt.Sprintf("SELECT firstMessage FROM topics WHERE uuid = '%s'", uuidPath)
+			row, err := db.Query(query2)
+			if err != nil {
+				fmt.Println(err)
 			} else {
-				fmt.Println("you need to be the owner of the topic to delete it")
+				for row.Next() {
+					defer row.Close()
+					err = row.Scan(&firstMessage)
+					if err != nil {
+						fmt.Println(err)
+					}
+				}
+			}
+
+			fmt.Println(firstMessage, message)
+
+			if firstMessage == message {
+				fmt.Println("you can't delete the first message of a topic")
+			} else {
+				if owner == t.USER.Username || t.USER.Admin == 1 {
+					query2 := fmt.Sprintf("DELETE FROM messages WHERE uuid = '%s'", r.FormValue("delete"))
+					db.Exec(query2)
+					query = fmt.Sprintf("DELETE FROM likesFromUser WHERE uuidLiked = '%s'", r.FormValue("delete"))
+					db.Exec(query)
+					fmt.Println("message deleted")
+					query3 := fmt.Sprintf("UPDATE topics SET nmbPosts = nmbPosts - 1 WHERE uuid = '%s'", uuidPath)
+					db.Exec(query3)
+
+				} else {
+					fmt.Println("you need to be the owner of the topic to delete it or cant delete first message of topic, delete topic instead")
+				}
 			}
 		}
 	}
